agregado de parametros de entrada para todas las propeidades float que no me ponia la coma decimal. Falta hacer que haga bien las cuentas desde javascript cuando tiene punto o tiene coma

+++ SUA/Models/Bordereaux.cs
@@ -9,27 +9,27 @@ namespace SUA.Models
     {
         public List<EntradasBorderaux> Entradas { get; set; }
         public int EntradasTotal { get; set; }
-        public double EntradasBruto { get; set; }
+        public float EntradasBruto { get; set; }
 
         public List<ImpuestosDeduccionesTeatroBorderaux> ImpuestosDeduccionesTeatro { get; set; }
-        public double ImpuestosDeduccionesBruto { get; set; }
-        public double ImpuestosDeduccionesTotalDeducir { get; set; }
-        public double ImpuestosDeduccionesNeto { get; set; }
+        public float ImpuestosDeduccionesBruto { get; set; }
+        public float ImpuestosDeduccionesTotalDeducir { get; set; }
+        public float ImpuestosDeduccionesNeto { get; set; }
 
-        public double ImpuestosDeduccionesCompanyPorcentaje { get; set; }
-        public double ImpuestosDeduccionesCompanyMonto { get; set; }
-        public double ImpuestosDeduccionesTeatroPorcentaje { get; set; }
-        public double ImpuestosDeduccionesTeatroMonto { get; set; }
+        public float ImpuestosDeduccionesCompanyPorcentaje { get; set; }
+        public float ImpuestosDeduccionesCompanyMonto { get; set; }
+        public float ImpuestosDeduccionesTeatroPorcentaje { get; set; }
+        public float ImpuestosDeduccionesTeatroMonto { get; set; }
 
         public bool ArregloFijo { get; set; }
         public Dictionary<string, double> GastosCompany { get; set; }
-        public double GastosCompanyTotal { get; set; }
-        public double GastosCompanyNeto { get; set; }
+        public float GastosCompanyTotal { get; set; }
+        public float GastosCompanyNeto { get; set; }
 
-        public double SUAPorcentaje { get; set; }
-        public double SUAMonto { get; set; }
-        public double ShowPorcentaje { get; set; }
-        public double ShowMonto { get; set; }
+        public float SUAPorcentaje { get; set; }
+        public float SUAMonto { get; set; }
+        public float ShowPorcentaje { get; set; }
+        public float ShowMonto { get; set; }
 
         public DateTime FechaCobro { get; set; }
         public string Forma { get; set; }
